backend/requirements.txt
fastapi
uvicorn
numpy
Pillow
tensorflow
python-multipart
pytest